Update README.md
Guess-My-Number
Description:  This simple yet engaging game challenges players to guess a randomly generated number within a specified range. It's a fun project.
Firstly we have to enter a number it checks that it low or high of that unknown number. Then we have to guess the unknown number. It is quite Interesting.
Objective:  The theme of this game is kind of retro design.  In this simply we have to guess the number from 1 to 20.
How to play:
•	To play the game, simply run the program and follow the prompts to input your guesses. The game will provide feedback on each guess, indicating whether the correct number is higher or lower than the guess. Keep guessing until you find the correct number!
Features:
•	Randomly generated number within a specified range.
•	User-friendly interface for input guesses.
•	Feedback on each guess to help players define their guesses.
•	Built-in validation to ensure input is within the valid range.